hv: fix MISRA-C violations in vpci code: implicit conversion
458S: Implicit conversion: actual to formal param (MR): "The value of an
actual parameter shall not be implicitly converted to the type of a formal
parameter, if that conversion might result in the loss of information".

Tracked-On: #861

+++ hypervisor/dm/hw/pci.c
@@ -54,18 +54,18 @@ uint32_t pci_pdev_read_cfg(union pci_bdf bdf, uint32_t offset, uint32_t bytes)
 	addr = pci_pdev_calc_address(bdf, offset);
 
 	/* Write address to ADDRESS register */
-	pio_write32(addr, PCI_CONFIG_ADDR);
+	pio_write32(addr, (uint16_t)PCI_CONFIG_ADDR);
 
 	/* Read result from DATA register */
 	switch (bytes) {
 	case 1U:
-		val = (uint32_t)pio_read8(PCI_CONFIG_DATA + ((uint16_t)offset & 3U));
+		val = (uint32_t)pio_read8((uint16_t)PCI_CONFIG_DATA + ((uint16_t)offset & 3U));
 		break;
 	case 2U:
-		val = (uint32_t)pio_read16(PCI_CONFIG_DATA + ((uint16_t)offset & 2U));
+		val = (uint32_t)pio_read16((uint16_t)PCI_CONFIG_DATA + ((uint16_t)offset & 2U));
 		break;
 	default:
-		val = pio_read32(PCI_CONFIG_DATA);
+		val = pio_read32((uint16_t)PCI_CONFIG_DATA);
 		break;
 	}
 	spinlock_release(&pci_device_lock);
@@ -82,18 +82,18 @@ void pci_pdev_write_cfg(union pci_bdf bdf, uint32_t offset, uint32_t bytes,	uint
 	addr = pci_pdev_calc_address(bdf, offset);
 
 	/* Write address to ADDRESS register */
-	pio_write32(addr, PCI_CONFIG_ADDR);
+	pio_write32(addr, (uint16_t)PCI_CONFIG_ADDR);
 
 	/* Write value to DATA register */
 	switch (bytes) {
 	case 1U:
-		pio_write8((uint8_t)val, PCI_CONFIG_DATA + ((uint16_t)offset & 3U));
+		pio_write8((uint8_t)val, (uint16_t)PCI_CONFIG_DATA + ((uint16_t)offset & 3U));
 		break;
 	case 2U:
-		pio_write16((uint16_t)val, PCI_CONFIG_DATA + ((uint16_t)offset & 2U));
+		pio_write16((uint16_t)val, (uint16_t)PCI_CONFIG_DATA + ((uint16_t)offset & 2U));
 		break;
 	default:
-		pio_write32(val, PCI_CONFIG_DATA);
+		pio_write32(val, (uint16_t)PCI_CONFIG_DATA);
 		break;
 	}
 	spinlock_release(&pci_device_lock);
